dev-requirements.txt
django-extensions==3.2.0
django-debug-toolbar
-r requirements.txt